+++ src/zbc_ansi_internal.h
@@ -98,7 +98,7 @@ int zbc_ansi_tickfreq(void);
  *========================================================================*/
 
 int zbc_ansi_get_cmdline(char *buf, size_t buf_size);
-int zbc_ansi_heapinfo(uint64_t *heap_base, uint64_t *heap_limit,
-                      uint64_t *stack_base, uint64_t *stack_limit);
+int zbc_ansi_heapinfo(uintptr_t *heap_base, uintptr_t *heap_limit,
+                      uintptr_t *stack_base, uintptr_t *stack_limit);
 
 #endif /* ZBC_ANSI_INTERNAL_H */
